+++ integration/simple/simple.ts
@@ -1035,48 +1035,54 @@ export const SimpleWithMap = {
 
   fromJSON(object: any): SimpleWithMap {
     const message = { ...baseSimpleWithMap } as SimpleWithMap;
-    message.entitiesById = {};
-    if (object.entitiesById !== undefined && object.entitiesById !== null) {
-      Object.entries(object.entitiesById).forEach(([key, value]) => {
-        message.entitiesById[Number(key)] = Entity.fromJSON(value);
-      });
-    }
-    message.nameLookup = {};
-    if (object.nameLookup !== undefined && object.nameLookup !== null) {
-      Object.entries(object.nameLookup).forEach(([key, value]) => {
-        message.nameLookup[key] = String(value);
-      });
-    }
-    message.intLookup = {};
-    if (object.intLookup !== undefined && object.intLookup !== null) {
-      Object.entries(object.intLookup).forEach(([key, value]) => {
-        message.intLookup[Number(key)] = Number(value);
-      });
-    }
-    message.mapOfTimestamps = {};
-    if (object.mapOfTimestamps !== undefined && object.mapOfTimestamps !== null) {
-      Object.entries(object.mapOfTimestamps).forEach(([key, value]) => {
-        message.mapOfTimestamps[key] = fromJsonTimestamp(value);
-      });
-    }
-    message.mapOfBytes = {};
-    if (object.mapOfBytes !== undefined && object.mapOfBytes !== null) {
-      Object.entries(object.mapOfBytes).forEach(([key, value]) => {
-        message.mapOfBytes[key] = bytesFromBase64(value as string);
-      });
-    }
-    message.mapOfStringValues = {};
-    if (object.mapOfStringValues !== undefined && object.mapOfStringValues !== null) {
-      Object.entries(object.mapOfStringValues).forEach(([key, value]) => {
-        message.mapOfStringValues[key] = value as string | undefined;
-      });
-    }
-    message.longLookup = {};
-    if (object.longLookup !== undefined && object.longLookup !== null) {
-      Object.entries(object.longLookup).forEach(([key, value]) => {
-        message.longLookup[Number(key)] = Number(value);
-      });
-    }
+    message.entitiesById = Object.entries(object.entitiesById ?? {}).reduce<{ [key: number]: Entity }>(
+      (acc, [key, value]) => {
+        acc[Number(key)] = Entity.fromJSON(value);
+        return acc;
+      },
+      {}
+    );
+    message.nameLookup = Object.entries(object.nameLookup ?? {}).reduce<{ [key: string]: string }>(
+      (acc, [key, value]) => {
+        acc[key] = String(value);
+        return acc;
+      },
+      {}
+    );
+    message.intLookup = Object.entries(object.intLookup ?? {}).reduce<{ [key: number]: number }>(
+      (acc, [key, value]) => {
+        acc[Number(key)] = Number(value);
+        return acc;
+      },
+      {}
+    );
+    message.mapOfTimestamps = Object.entries(object.mapOfTimestamps ?? {}).reduce<{ [key: string]: Date }>(
+      (acc, [key, value]) => {
+        acc[key] = fromJsonTimestamp(value);
+        return acc;
+      },
+      {}
+    );
+    message.mapOfBytes = Object.entries(object.mapOfBytes ?? {}).reduce<{ [key: string]: Uint8Array }>(
+      (acc, [key, value]) => {
+        acc[key] = bytesFromBase64(value as string);
+        return acc;
+      },
+      {}
+    );
+    message.mapOfStringValues = Object.entries(object.mapOfStringValues ?? {}).reduce<{
+      [key: string]: string | undefined;
+    }>((acc, [key, value]) => {
+      acc[key] = value as string | undefined;
+      return acc;
+    }, {});
+    message.longLookup = Object.entries(object.longLookup ?? {}).reduce<{ [key: number]: number }>(
+      (acc, [key, value]) => {
+        acc[Number(key)] = Number(value);
+        return acc;
+      },
+      {}
+    );
     return message;
   },
 
@@ -1129,62 +1135,68 @@ export const SimpleWithMap = {
 
   fromPartial(object: DeepPartial<SimpleWithMap>): SimpleWithMap {
     const message = { ...baseSimpleWithMap } as SimpleWithMap;
-    message.entitiesById = {};
-    if (object.entitiesById !== undefined && object.entitiesById !== null) {
-      Object.entries(object.entitiesById).forEach(([key, value]) => {
+    message.entitiesById = Object.entries(object.entitiesById ?? {}).reduce<{ [key: number]: Entity }>(
+      (acc, [key, value]) => {
         if (value !== undefined) {
-          message.entitiesById[Number(key)] = Entity.fromPartial(value);
+          acc[Number(key)] = Entity.fromPartial(value);
         }
-      });
-    }
-    message.nameLookup = {};
-    if (object.nameLookup !== undefined && object.nameLookup !== null) {
-      Object.entries(object.nameLookup).forEach(([key, value]) => {
+        return acc;
+      },
+      {}
+    );
+    message.nameLookup = Object.entries(object.nameLookup ?? {}).reduce<{ [key: string]: string }>(
+      (acc, [key, value]) => {
         if (value !== undefined) {
-          message.nameLookup[key] = String(value);
+          acc[key] = String(value);
         }
-      });
-    }
-    message.intLookup = {};
-    if (object.intLookup !== undefined && object.intLookup !== null) {
-      Object.entries(object.intLookup).forEach(([key, value]) => {
+        return acc;
+      },
+      {}
+    );
+    message.intLookup = Object.entries(object.intLookup ?? {}).reduce<{ [key: number]: number }>(
+      (acc, [key, value]) => {
         if (value !== undefined) {
-          message.intLookup[Number(key)] = Number(value);
+          acc[Number(key)] = Number(value);
         }
-      });
-    }
-    message.mapOfTimestamps = {};
-    if (object.mapOfTimestamps !== undefined && object.mapOfTimestamps !== null) {
-      Object.entries(object.mapOfTimestamps).forEach(([key, value]) => {
+        return acc;
+      },
+      {}
+    );
+    message.mapOfTimestamps = Object.entries(object.mapOfTimestamps ?? {}).reduce<{ [key: string]: Date }>(
+      (acc, [key, value]) => {
         if (value !== undefined) {
-          message.mapOfTimestamps[key] = value;
+          acc[key] = value;
         }
-      });
-    }
-    message.mapOfBytes = {};
-    if (object.mapOfBytes !== undefined && object.mapOfBytes !== null) {
-      Object.entries(object.mapOfBytes).forEach(([key, value]) => {
+        return acc;
+      },
+      {}
+    );
+    message.mapOfBytes = Object.entries(object.mapOfBytes ?? {}).reduce<{ [key: string]: Uint8Array }>(
+      (acc, [key, value]) => {
         if (value !== undefined) {
-          message.mapOfBytes[key] = value;
+          acc[key] = value;
         }
-      });
-    }
-    message.mapOfStringValues = {};
-    if (object.mapOfStringValues !== undefined && object.mapOfStringValues !== null) {
-      Object.entries(object.mapOfStringValues).forEach(([key, value]) => {
+        return acc;
+      },
+      {}
+    );
+    message.mapOfStringValues = Object.entries(object.mapOfStringValues ?? {}).reduce<{
+      [key: string]: string | undefined;
+    }>((acc, [key, value]) => {
+      if (value !== undefined) {
+        acc[key] = value;
+      }
+      return acc;
+    }, {});
+    message.longLookup = Object.entries(object.longLookup ?? {}).reduce<{ [key: number]: number }>(
+      (acc, [key, value]) => {
         if (value !== undefined) {
-          message.mapOfStringValues[key] = value;
+          acc[Number(key)] = Number(value);
         }
-      });
-    }
-    message.longLookup = {};
-    if (object.longLookup !== undefined && object.longLookup !== null) {
-      Object.entries(object.longLookup).forEach(([key, value]) => {
-        if (value !== undefined) {
-          message.longLookup[Number(key)] = Number(value);
-        }
-      });
-    }
+        return acc;
+      },
+      {}
+    );
     return message;
   },
 };
@@ -1618,12 +1630,13 @@ export const SimpleWithSnakeCaseMap = {
 
   fromJSON(object: any): SimpleWithSnakeCaseMap {
     const message = { ...baseSimpleWithSnakeCaseMap } as SimpleWithSnakeCaseMap;
-    message.entitiesById = {};
-    if (object.entitiesById !== undefined && object.entitiesById !== null) {
-      Object.entries(object.entitiesById).forEach(([key, value]) => {
-        message.entitiesById[Number(key)] = Entity.fromJSON(value);
-      });
-    }
+    message.entitiesById = Object.entries(object.entitiesById ?? {}).reduce<{ [key: number]: Entity }>(
+      (acc, [key, value]) => {
+        acc[Number(key)] = Entity.fromJSON(value);
+        return acc;
+      },
+      {}
+    );
     return message;
   },
 
@@ -1640,14 +1653,15 @@ export const SimpleWithSnakeCaseMap = {
 
   fromPartial(object: DeepPartial<SimpleWithSnakeCaseMap>): SimpleWithSnakeCaseMap {
     const message = { ...baseSimpleWithSnakeCaseMap } as SimpleWithSnakeCaseMap;
-    message.entitiesById = {};
-    if (object.entitiesById !== undefined && object.entitiesById !== null) {
-      Object.entries(object.entitiesById).forEach(([key, value]) => {
+    message.entitiesById = Object.entries(object.entitiesById ?? {}).reduce<{ [key: number]: Entity }>(
+      (acc, [key, value]) => {
         if (value !== undefined) {
-          message.entitiesById[Number(key)] = Entity.fromPartial(value);
+          acc[Number(key)] = Entity.fromPartial(value);
         }
-      });
-    }
+        return acc;
+      },
+      {}
+    );
     return message;
   },
 };
@@ -1742,12 +1756,13 @@ export const SimpleWithMapOfEnums = {
 
   fromJSON(object: any): SimpleWithMapOfEnums {
     const message = { ...baseSimpleWithMapOfEnums } as SimpleWithMapOfEnums;
-    message.enumsById = {};
-    if (object.enumsById !== undefined && object.enumsById !== null) {
-      Object.entries(object.enumsById).forEach(([key, value]) => {
-        message.enumsById[Number(key)] = value as number;
-      });
-    }
+    message.enumsById = Object.entries(object.enumsById ?? {}).reduce<{ [key: number]: StateEnum }>(
+      (acc, [key, value]) => {
+        acc[Number(key)] = value as number;
+        return acc;
+      },
+      {}
+    );
     return message;
   },
 
@@ -1764,14 +1779,15 @@ export const SimpleWithMapOfEnums = {
 
   fromPartial(object: DeepPartial<SimpleWithMapOfEnums>): SimpleWithMapOfEnums {
     const message = { ...baseSimpleWithMapOfEnums } as SimpleWithMapOfEnums;
-    message.enumsById = {};
-    if (object.enumsById !== undefined && object.enumsById !== null) {
-      Object.entries(object.enumsById).forEach(([key, value]) => {
+    message.enumsById = Object.entries(object.enumsById ?? {}).reduce<{ [key: number]: StateEnum }>(
+      (acc, [key, value]) => {
         if (value !== undefined) {
-          message.enumsById[Number(key)] = value as number;
+          acc[Number(key)] = value as number;
         }
-      });
-    }
+        return acc;
+      },
+      {}
+    );
     return message;
   },
 };